Migrate cli implementation to new `app` module
