+++ src/app/cli.rs
